+++ lib/blocs/storage/blocs.dart
@@ -4,5 +4,4 @@ export 'item_form/item_form_bloc.dart';
 export 'search/search_bloc.dart';
 export 'storage_detail/storage_detail_bloc.dart';
 export 'storage_edit/storage_edit_bloc.dart';
-export 'storage_form/storage_form_bloc.dart';
 export 'storage_home/storage_home_bloc.dart';
